Refactor code into smaller pieces
- Move all classes to their own files
- Pass Date objects around rather than formatted strings.
  Different APIs need different string formats
